requirements.txt
asyncio
aiohttp
discord.py
